+++ CHANGES.md
@@ -1,9 +1,17 @@
 # Changes Logs
 
+## v2.1.0
+
+- Added encoding `base32`.
+
+## v2.0.0
+
+- Simplified the APIs.
+
 ## v1.1.1
 
 - Fixed the `convert` result from `base62x` to `buffer`, `hex`, `base64`, `utf8`.
-- Improved the sample `04-compare` with **binary-safe** tests.
+- Improved the example `04-compare` with **binary-safe** tests.
 - Added supports for encoding `urlencode`.
 
 ## v1.1.0